# artifacts/api-server/src/routes/search.ts
import { Router, type IRouter } from "express";
import { desc, ilike, or, sql } from "drizzle-orm";
import { db, documentsTable, searchHistoryTable } from "@workspace/db";
import {
  SemanticSearchBody,
  GetSearchHistoryQueryParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

function docToResponse(doc: typeof documentsTable.$inferSelect) {
  return {
    id: String(doc.id),
    title: doc.title,
    path: doc.path,
    type: doc.type,
    size: doc.size,
    summary: doc.summary ?? null,
    keywords: doc.keywords ?? [],
    preview: doc.preview ?? null,
    ocrText: doc.ocrText ?? null,
    indexed: doc.indexed,
    embeddingCount: doc.embeddingCount ?? null,
    createdAt: doc.createdAt.toISOString(),
    modifiedAt: doc.modifiedAt.toISOString(),
  };
}

router.post("/search", async (req, res): Promise<void> => {
  const start = Date.now();
  const parsed = SemanticSearchBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { query, limit = 10, types } = parsed.data;

  // In production this uses nomic-embed-text + ChromaDB vector search.
  // Here we do keyword-based search as a functional demo.
  const queryWords = query
    .toLowerCase()
    .split(/\s+/)
    .filter((w) => w.length > 2);

  let dbQuery = db.select().from(documentsTable).$dynamic();

  if (queryWords.length > 0) {
    const conditions = queryWords.slice(0, 3).flatMap((word) => [
      ilike(documentsTable.title, `%${word}%`),
      ilike(documentsTable.preview, `%${word}%`),
      ilike(documentsTable.summary, `%${word}%`),
      ilike(documentsTable.ocrText, `%${word}%`),
    ]);
    // Use the first condition to ensure type safety
    if (conditions.length > 0) {
      dbQuery = dbQuery.where(or(...conditions));
    }
  }

  if (types && types.length > 0) {
    // filter by types - simplified for demo
    const allowedDocs = await dbQuery
      .orderBy(desc(documentsTable.modifiedAt))
      .limit(limit * 2);
    const typeFiltered = allowedDocs.filter((d) => types.includes(d.type));
    const results = typeFiltered.slice(0, limit).map((doc) => ({
      document: docToResponse(doc),
      score: Math.random() * 0.4 + 0.6,
      snippet:
        doc.preview?.slice(0, 200) ??
        `Relevant content found in ${doc.title} matching your search for "${query}".`,
    }));
    const elapsed = Date.now() - start;

    await db.insert(searchHistoryTable).values({
      query,
      resultCount: results.length,
    });

    res.json({
      query,
      results,
      totalResults: results.length,
      aiResponse: generateAiResponse(query, results.length),
      timeTakenMs: elapsed,
    });
    return;
  }

  const docs = await dbQuery
    .orderBy(desc(documentsTable.modifiedAt))
    .limit(limit);

  // Assign similarity scores (mock semantic scores)
  const results = docs.map((doc) => {
    const titleMatch = doc.title.toLowerCase().includes(query.toLowerCase())
      ? 0.3
      : 0;
    const baseScore = 0.5 + titleMatch + Math.random() * 0.2;
    return {
      document: docToResponse(doc),
      score: Math.min(baseScore, 0.99),
      snippet:
        doc.preview?.slice(0, 200) ??
        `Relevant content found in ${doc.title} matching your search for "${query}".`,
    };
  });

  // Sort by score descending
  results.sort((a, b) => b.score - a.score);

  const elapsed = Date.now() - start;

  await db.insert(searchHistoryTable).values({
    query,
    resultCount: results.length,
  });

  res.json({
    query,
    results,
    totalResults: results.length,
    aiResponse: generateAiResponse(query, results.length),
    timeTakenMs: elapsed,
  });
});

function generateAiResponse(query: string, count: number): string {
  if (count === 0) {
    return `I couldn't find any documents matching "${query}". Try indexing more folders in Settings, or rephrase your query.`;
  }
  return `I found ${count} document${count === 1 ? "" : "s"} related to "${query}". The most relevant results are shown below, ranked by semantic similarity. In production mode, this response would be generated by Gemma 3 running locally via Ollama.`;
}

router.get("/search/history", async (req, res): Promise<void> => {
  const parsed = GetSearchHistoryQueryParams.safeParse(req.query);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const limit = parsed.data.limit ?? 20;
  const history = await db
    .select()
    .from(searchHistoryTable)
    .orderBy(desc(searchHistoryTable.timestamp))
    .limit(limit);

  res.json({
    history: history.map((h) => ({
      id: String(h.id),
      query: h.query,
      timestamp: h.timestamp.toISOString(),
      resultCount: h.resultCount,
    })),
  });
});

export default router;