# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MemoryOS API - AI-powered semantic document search and indexing
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentType = typeof DocumentType[keyof typeof DocumentType];


export const DocumentType = {
  pdf: 'pdf',
  docx: 'docx',
  txt: 'txt',
  md: 'md',
  image: 'image',
  unknown: 'unknown',
} as const;

export interface Document {
  id: string;
  title: string;
  path: string;
  type: DocumentType;
  /** File size in bytes */
  size: number;
  /** @nullable */
  summary?: string | null;
  keywords?: string[];
  /**
     * Text preview snippet
     * @nullable
     */
  preview?: string | null;
  /**
     * OCR-extracted text for images
     * @nullable
     */
  ocrText?: string | null;
  createdAt: string;
  modifiedAt: string;
  indexed: boolean;
  /** @nullable */
  embeddingCount?: number | null;
}

export interface DocumentList {
  documents: Document[];
  total: number;
}

export interface DocumentSummary {
  id: string;
  summary: string;
  keywords: string[];
  keyPoints: string[];
}

export interface SearchInput {
  /** @minLength 1 */
  query: string;
  limit?: number;
  types?: string[];
}

export interface SearchResultItem {
  document: Document;
  /** Similarity score 0-1 */
  score: number;
  /** Relevant excerpt from document */
  snippet: string;
}

export interface SearchResult {
  query: string;
  results: SearchResultItem[];
  totalResults: number;
  /**
     * LLM-generated natural language response
     * @nullable
     */
  aiResponse: string | null;
  timeTakenMs: number;
}

export interface SearchHistory {
  id: string;
  query: string;
  timestamp: string;
  resultCount: number;
}

export interface SearchHistoryList {
  history: SearchHistory[];
}

export interface IndexInput {
  /** Absolute folder path to index */
  path: string;
  recursive?: boolean;
}

export interface IndexStatus {
  isIndexing: boolean;
  totalFiles: number;
  processedFiles: number;
  failedFiles: number;
  /** @nullable */
  currentFile: string | null;
  folders: string[];
}

export interface OcrInput {
  filePath: string;
}

export interface OcrResult {
  filePath: string;
  text: string;
  confidence: number;
}

export type DashboardStatsTypeBreakdown = {
  pdf: number;
  docx: number;
  txt: number;
  md: number;
  image: number;
};

export interface DayActivity {
  date: string;
  indexed: number;
  searched: number;
}

export interface DashboardStats {
  totalDocuments: number;
  totalImages: number;
  totalEmbeddings: number;
  storageUsedBytes: number;
  recentSearches: number;
  typeBreakdown: DashboardStatsTypeBreakdown;
  indexedToday: number;
  activityByDay: DayActivity[];
}

export interface TimelineGroup {
  label: string;
  date: string;
  documents: Document[];
  searches: SearchHistory[];
}

export interface Timeline {
  groups: TimelineGroup[];
}

export interface Folder {
  path: string;
  documentCount: number;
  /** @nullable */
  lastIndexed: string | null;
  isWatched: boolean;
}

export interface FolderList {
  folders: Folder[];
}

export type ListDocumentsParams = {
type?: ListDocumentsType;
limit?: number;
offset?: number;
};

export type ListDocumentsType = typeof ListDocumentsType[keyof typeof ListDocumentsType];


export const ListDocumentsType = {
  pdf: 'pdf',
  docx: 'docx',
  txt: 'txt',
  md: 'md',
  image: 'image',
  all: 'all',
} as const;

export type GetSearchHistoryParams = {
limit?: number;
};

export type GetTimelineParams = {
days?: number;
};