# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MemoryOS API - AI-powered semantic document search and indexing
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentType } from './documentType';

export interface Document {
  id: string;
  title: string;
  path: string;
  type: DocumentType;
  /** File size in bytes */
  size: number;
  /** @nullable */
  summary?: string | null;
  keywords?: string[];
  /**
     * Text preview snippet
     * @nullable
     */
  preview?: string | null;
  /**
     * OCR-extracted text for images
     * @nullable
     */
  ocrText?: string | null;
  createdAt: Date;
  modifiedAt: Date;
  indexed: boolean;
  /** @nullable */
  embeddingCount?: number | null;
}