# lib/api-zod/src/generated/types/documentList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MemoryOS API - AI-powered semantic document search and indexing
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface DocumentList {
  documents: Document[];
  total: number;
}