# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MemoryOS API - AI-powered semantic document search and indexing
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}